+++ src/main/java/tn/dalhia/services/implementations/SubscriptionServiceImpl.java
@@ -10,6 +10,7 @@ import java.util.List;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.scheduling.annotation.Scheduled;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -47,9 +48,11 @@ public class SubscriptionServiceImpl implements SubscriptionService  {
 
 	@Override
 	@Transactional
-	public SubscriptionDto createSubscription(SubscriptionRequestModel subscription) {
+	public SubscriptionDto createSubscription(SubscriptionRequestModel subscription, Authentication authentification) {
 		User userEntity = userRepo.findByUserId(subscription.getUserId());
 		Plan plan = planRepo.findById(subscription.getPlanId()).orElse(null);
+		
+		if(!utils.connectedUser(authentification,userEntity)) throw new UserServiceException(ErrorMessages.SECURITY_ERROR.getErrorMessage());
 		if (userEntity == null) throw new UserServiceException(ErrorMessages.NO_RECORD_FOUND.getErrorMessage());
 		if (plan == null) throw new UserServiceException(ErrorMessages.NO_RECORD_FOUND.getErrorMessage());
 		
@@ -71,10 +74,12 @@ public class SubscriptionServiceImpl implements SubscriptionService  {
 	}
 
 	@Override
-	public SubscriptionDto updateSubscription(SubscriptionRequestModel subscription, String id) {
+	public SubscriptionDto updateSubscription(SubscriptionRequestModel subscription, String id, Authentication authentification) {
+		if(!utils.connectedUser(authentification,null)) throw new UserServiceException(ErrorMessages.SECURITY_ERROR.getErrorMessage());
 		SubscriptionDto returnValue = new SubscriptionDto();
 		Subscription sub = subscriptionRepo.findBySubscritpionId(id);
 		if (sub == null) throw new UserServiceException(ErrorMessages.NO_RECORD_FOUND.getErrorMessage());
+		
 		
 		sub.setDate_debut(subscription.getDate_debut());
 		sub.setDate_fin(subscription.getDate_fin());
@@ -86,9 +91,12 @@ public class SubscriptionServiceImpl implements SubscriptionService  {
 	}
 
 	@Override
-	public SubscriptionDto getSubscriptionById(String id) {
+	public SubscriptionDto getSubscriptionById(String id, Authentication authentification) {
 		SubscriptionDto returnValue= new SubscriptionDto();
 		Subscription sub = subscriptionRepo.findBySubscritpionId(id);
+		
+		if (sub == null) throw new UserServiceException(ErrorMessages.NO_RECORD_FOUND.getErrorMessage());
+		if(!utils.connectedUser(authentification,sub.getUserDetails())) throw new UserServiceException(ErrorMessages.SECURITY_ERROR.getErrorMessage());
 		
 		BeanUtils.copyProperties(sub,returnValue);
 		returnValue.setMessage(RequestOperationStatus.SUCCESS.name());
@@ -97,10 +105,12 @@ public class SubscriptionServiceImpl implements SubscriptionService  {
 	}
 
 	@Override
-	public void deleteSubscription(String id) {
-		Subscription sub = subscriptionRepo.findBySubscritpionId(id);
-		if (sub == null) throw new UserServiceException(ErrorMessages.NO_RECORD_FOUND.getErrorMessage());
+	public void deleteSubscription(String id, Authentication authentification) {
 		
+		Subscription sub = subscriptionRepo.findBySubscritpionId(id);
+		
+		if (sub == null) throw new UserServiceException(ErrorMessages.NO_RECORD_FOUND.getErrorMessage());
+		if(!utils.connectedUser(authentification,sub.getUserDetails())) throw new UserServiceException(ErrorMessages.SECURITY_ERROR.getErrorMessage());
 		User user = userRepo.findBySubscriptionsId(sub.getId());
 		if (user == null) {
 		subscriptionRepo.delete(sub);
